refactor(src/common/api/AuthSvc): move functions that manipulates with cookies and localStorage to own LocalStorageCookies service
